Rename hardware keyrings file and small review fixes
