+++ components/brave_wallet_ui/common/api/hardware_keyrings.ts
