+++ components/brave_wallet_ui/common/async/lib.ts
@@ -21,7 +21,7 @@ import getAPIProxy from './bridge'
 import { Dispatch, State } from './types'
 import LedgerBridgeKeyring from '../../common/hardware/ledgerjs/eth_ledger_bridge_keyring'
 import TrezorBridgeKeyring from '../../common/hardware/trezor/trezor_bridge_keyring'
-import { getHardwareKeyring } from '../api/getKeyringsByType'
+import { getHardwareKeyring } from '../api/hardware_keyrings'
 import { HardwareWalletAccount } from '../hardware/types'
 import { GetAccountsHardwareOperationResult } from '../hardware_operations'
